alteração no grupo e nome do projeto
