+++ src/main/java/br/com/cattle_control/starter/infra/converter/EntityConverter.java
@@ -1,4 +1,4 @@
-package com.github.adminfaces.starter.infra.converter;
+package br.com.cattle_control.starter.infra.converter;
 
 
 import javax.faces.component.UIComponent;
