+++ src/main/webapp/WEB-INF/templates/template.xhtml
@@ -6,7 +6,7 @@
                 template="/admin.xhtml">
 
     <ui:define name="head">
-        <title>Admin Starter</title>
+        <title>Cattle Control</title>
         <link rel="shortcut icon" type="image/x-icon" href="#{resource['favicon/favicon.ico']}"/>
         <link rel="shortcut icon" type="image/x-icon" href="#{resource['favicon/favicon-16x16.png']}" sizes="16x16"/>
         <link rel="shortcut icon" type="image/x-icon" href="#{resource['favicon/favicon-32x32.png']}" sizes="32x32"/>
@@ -18,7 +18,7 @@
     </ui:define>
 
     <ui:define name="logo">
-        Admin Starter
+        Cattle Control
     </ui:define>
 
     <ui:define name="logo-mini">
