+++ src/main/webapp/index.xhtml
@@ -10,15 +10,11 @@
 
     <ui:define name="title">
         <h2 class="align-center">
-            Welcome to the <span class="text-aqua"> <i><a href="https://github.com/adminfaces/admin-starter-springboot"
+            <span class="text-aqua"> <i><a href="https://github.com/hmarcuzzo/cattle_control"
                                                           target="_blank"
-                                                          style="text-transform: none;text-decoration: none"> AdminFaces SpringBoot Starter</a></i></span> Project!
+                                                          style="text-transform: none;text-decoration: none"> Cattle Control</a></i></span>
             <br/>
-            <small>Integrating <p:link value="Primefaces" href="http://primefaces.org"/>, <p:link value="Bootstrap"
-                                                                                                  href="http://getbootstrap.com/"/> and
-                <p:link value="Admin LTE" href="https://almsaeedstudio.com/themes/AdminLTE/index2.html/"/> into your
-                <p:link value="JSF" href="https://javaserverfaces.java.net/"/> application.
-            </small>
+            <small>Software de gerenciamento destinado à pecuárias para o auxílio no controle zootécnico e financeiro de gado de corte.</small>
         </h2>
 
     </ui:define>
